p1/requirements.txt
pyspark
delta-spark==2.4.0
pandas
openpyxl
minio
dotenv
boto3
